optimize PageReader and ChunkReader
